refactor: reduce some unneed module
