+++ packages/vrender-core/src/animate/Ticker/index.ts
@@ -1,1 +1,5 @@
 export * from './default-ticker';
+export * from './manual-ticker';
+export * from './manual-ticker-handler';
+export * from './raf-tick-handler';
+export * from './timeout-tick-handler';
